+++ src/app/mypage/components/Common/index.tsx
@@ -1,3 +1,3 @@
 export { default as FileThumbnails } from './FileThumbnails';
-export { default as useDefaultFile } from './useDefaultFile';
+export { default as useReadFile } from './useReadFile';
 export { default as FileEdit } from './FileEdit';
