+++ src/app/mypage/components/Common/useReadFile.tsx
